Move setVolume function into volume folder
